Added basic support for test coverage.
Blanket is a simple add-on for Mocha which allows coverage to be calculated for tests. The reporters are separate right now, but in the future when mocha can support multiple reporters both test and coverage results can be output.

Removed old versions of node (0.6,0.8) since they are not supported by the coverage reporter.